now onto clear completed
