--- src/Todo.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default function Todo (props){
-  return <p>I'm a little taskbot short and sweet</p>
-}
