+++ src/components/Todo.css
@@ -1,0 +1,6 @@
+p{
+  font-weight:bold;
+  margin: 5px 0;
+  font-size:2rem;
+  color:rgb(0, 121, 36);
+}
