+++ src/components/Todo.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import './Todo.css'
+
+export default function Todo (props){
+  return (
+
+  <p id={props.id}>{props.task}</p>)
+}
